requirements.txt
aniso8601==7.0.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.10.2
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.1
pluggy==0.12.0
promise==2.2.1
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
Rx==1.6.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
whitenoise==4.1.3
wrapt==1.11.2
zipp==0.5.2
